Patient Delete

+++ frontend/src/components/PatientDetails.jsx
@@ -1,10 +1,12 @@
-import { Link } from 'react-router-dom';
+import { Link, useNavigate } from 'react-router-dom';
 import { useState, useEffect } from "react";
 import { useAuth } from '../context/AuthContext';
 import axiosInstance from "../axiosConfig";
 
 const PatientDetails = ({ patients, setPatients, setEditingPatient }) => {
     const { user } = useAuth();
+    const [showConfirm, setShowConfirm] = useState(false);
+    const navigate = useNavigate();
 
     const handleDelete = async (patientID) => {
         try {
@@ -12,9 +14,12 @@ const PatientDetails = ({ patients, setPatients, setEditingPatient }) => {
                 headers: { Authorization: `Bearer ${user.token}` },
             });
             alert('Success! Patient deleted.')
-            setPatients(patients.filter((patient) => patient._id !== patientID));
+            navigate('/patients');
+            // setPatients(patients.filter((patient) => patient._id !== patientID));
         } catch (error) {
             alert('Error: Failed to delete patient.')
+        } finally {
+            setShowConfirm(false);
         }
     };
 
@@ -67,34 +72,61 @@ const PatientDetails = ({ patients, setPatients, setEditingPatient }) => {
                     </div>
 
                     {/* row for history */}
-                    <div className="px-6">
+                    <div className="mt-2 p-6">
                         <label for="history">History:</label>
                         <textarea
                             id="history"
                             name="history"
                             type="text"
-                            className="h-80 w-full border rounded"
-                        // value=formData.history
+                            className="h-80 mb-4 p-2 w-full border rounded"
+                            // value={formData.history}
                         />
                     </div>
 
-                    {/* row buttons */}
-                    <div className="p-6">
+                    {/* row for buttons */}
+                    <div className="mt-2 p-6">
                         <Link to=""
                             className="pill-button bg-yellow-500 text-white px-4 py-2 rounded">
                             Edit
                         </Link>
                         <button
-                            onClick={() => handleDelete(patients._id)}
+                            // onClick={() => handleDelete(patients._id)}
+                            onClick={() => setShowConfirm(true)}
                             className="pill-button bg-red-500 text-white ml-2 px-4 py-2 rounded"
                         >
                             Delete
                         </button>
                     </div>
                 </div>
-                {/* ))} */}
             </div>
-        </div >
+
+            {/* show confirm */}
+            {showConfirm && (
+                <div className="fixed inset-0 flex items-center justify-center bg-black/50">
+                    <div className="bg-white p-6 rounded shadow text-center max-w-sm mx-4">
+                        <p className="mb-4 font-medium text-lg">
+                            Are you sure you want to delete patient profile?
+                        </p>
+
+                        <div className="flex justify-center gap-4">
+                            <button
+                                onClick={() => handleDelete(patients._id)}
+                                className="px-4 py-2 bg-red-600 text-white rounded-full"
+                            >
+                                Yes
+                            </button>
+
+                            <button
+                                onClick={() => setShowConfirm(false)}
+                                className="px-4 py-2 bg-gray-300 rounded-full"
+                            >
+                                No
+                            </button>
+                        </div>
+                    </div>
+                </div>
+            )}
+        </div>
     );
 };
